HPCC-9088 Python errors prevent start of ESP in 4.0.0rc2
Python library calls exit() if stdin is referring to a directory.

FOr some reason, this seems to sometimes happen if the R library is also
loaded. While we should also address the problem in the R embedding plugin, it
is never a good idea for the Python plugin to cause the caller to exit.
Therefore, we check for the situation that causes it to exit, and disable
the python support if it is present.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>