HPCC-9112 Large sets of files sprayed using wildcard may be truncated
This change modifies CRemoteDirectoryIterator::appendBuf() to properly
handle end of stream and end of block conditions. Before this fix if the
directory contents spanned more than one transfer buffer between nodes, only
the first buffer's contents would be iterated by next(), because the
block_end_flag was left in the buffer, and next() interpreted it to mean
'no_more_valid_items'.

This fix removes the block_end_flags.

It also changes a variable name from "b" to "isValidEntry" in the
CRemoteDirectoryIterator::next() method to improve code self-documentation.

Finally it changes some information text message in dfuplus.cpp because
the same log was generated in two different function (fixedSpray and
variableSpray).

Signed-off-by: Attila Vamos <attila.vamos@gmail.com>