HPCC-9189 Disable udpResendEnabled logic
Still seeing some issues even with the behaviour defaulted off, from people
that created environments while the default was still on.

It is definitely broken, so until I have a chance to work out why, disable it.
Allow it to be enabled only in debug builds (so that I can debug why it is
failing).

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>