HPCC-9033 Track number of ATMOST triggers in Roxie
Add a count of atmost triggers to the per-query stats.

The per-query stats can be accessed by adding timing='1' to the query XML
(adding log='1' is also usefuil, so that the timings are returned to the
caller rather than just appearing in the Roxie log.

e.g. testsocket . '<my_query log="1" timing="1"/>'

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>