HPCC-9062 Limit remote location list by cluster
Roxies with copyResources set to false will show up in the list of
possible remote file locations, but are not really suitable locations.

With an arbitrary maximum imposed on the list this can quickly fill
up the list with unlikely locations resulting in no file part being
found.

Limiting the list of possible remote file location entries by
cluster rather than an arbitrary maximum will prevent
copyResources=false roxies from monopolizing the list and
increase the likelyhood that a copy of a file part will
be found.

Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>