HPCC-9214 Include remoteSubFiles when replicating CResolvedFile on slave
Remote sub files were not being serialized/deserialized when
CResolvedFiles were copied to slaves as CSlaveDynamicFile.

subFiles and remoteSubFiles must have the same number of items or
issues will arise when invalid item numbers are accessed.  The remote
file descriptors were also not accessible on the slave nodes.

Also move a call to queryMatchingRemotePart to inside the if block where
the result is used.

Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>