HPCC-9225 Allow WsEcl caller to specify jobname
By default WsEcl gives a result workunit the same jobname as the
published query workunit.  Allow _jobname=<name> to be specified in
the url used to submit the query so that the user can create a
unique jobname each time the query is invoked.

Works for HTTP-GET URL, SOAP, XML POST, FORM POST, JSON POST.

Signed-off-by: Anthony Fishbeck <Anthony.Fishbeck@lexisnexis.com>