HPCC-9271 Show 'truncated' message for viewing large file
Inside the Helpers section of the ECL Workunit Details page, there is a
640k limit for displaying a cpp file, xml file, and workunit archive
file. This fix checks the size of the helper file. If it is larger than
640k bytes, a message '(Truncated to 640K bytes)' will be added beside
the View File link.

When a helper file is downloaded, the existing code uses 'file path
 + file name' as 'attachment filename'. Internet browser replaces
the Path Separators in the file path with the '_'s and the file
name looks strange. This fix removes the file path from the
'attachment filename'.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>