HPCC-9241 Returns results after action button is clicked on WUs page
On Browse (ECL or DFU) Workunits page, when an action (Protect,
SetToFailed, etc) button is clicked and the action is finished, the
page will be redirected to updated Browse Workunits page. That does
not work when rawxml_ flag is true (and for SOAP calls). This fix
returns action results if the rawxml_ flag or SOAP calls is detected.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>