HPCC-9306 Don't allow restarted workunits to take the wrong branch
This fix ensures that conditions for conditional workflow actions are always
re-evaluated.  Otherwise, the condition can evaluate to true, the true branch
fails.  The workunit is resubmitted using "recover", the condition wasn't
re-evaluated, causing the false branch to be taken the second time.

The same issue could occur if there was a failure in a persist, possibly leading
to an internal error, or the persist not being re-created correctly.

Signed-off-by: Gavin Halliday <gavin.halliday@lexisnexis.com>