HPCC-8379 Show suspended roxie query on EclWatch
If roxie suspends the query, querysets may not be updated and existing
ECLWatch page shows that it is not suspended. This fix calls roxie for
query state before displaying the suspended status. If a query is
suspended by a roxie, this fix shows that the query is suspended on
the cluster. If a query is suspended by a user, it shows that the
query is suspended by user. The fix works on both QuerySet page and
QueryDetails page.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>