HPCC-9369 Disable Delete button when a WU is running/aborting
A problem is reported when a user deleted an ECL workunit that was in
an "aborting" state. This fix disables the Delete button on both
Browse ECL WUs page and ECL WU Details page if the workunit is in one
of the following states: aborting, running, compiling, debug_running,
and debugging.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>