HPCC-9437 Add option for high resolution timing in log files
Change the meaning of the milli and micro logging options (no currently being
used by any client) so that the milli/microseconds are reported as a fraction
of the seconds in the time field (much easier to read).

Undo an unintentional change to the way that fields are selected in the Roxie
logs - it was supposed to (and used to) use a shorter set of fields when
logging to stdout, but during a precious refactoring was changed so that if
you log to stdout, the fields used for the logfile are restricted.

Add ability to configure logging format globally in environment.conf

Windows does not support the sub-second timings.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>