HPCC-9453 - keyed join, onfail causing record loss
Records transformed by onfail were being output, but it
broke the join group markers and caused a loss of records beyond
the ONFAIL'd record output.
There were separate but similar issues if preserving order of LHS,
which is the default, of if LHS grouping was preserved (default
if LHS is grouped)

The issue can be avoided by turning off record order preservation
and/or group order preservation.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>