HPCC-8515 Running standalone with dalisrvers=... has issues.
Currently when running a standalone workunit and specifying a DALI, a
workunit is created but is missing some information. Also, there is not
user credentials, so calls to dali pass a NULL user descriptor.
This fix updates the workunit with the missing Owner (which can now be
specified on the command line), job name ( which is argv[0]) and cluster
"StandAloneHThor".  I added a new "-USER=username:password" command line
argument, which will provide the credentials for DALI access.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>