HPCC-9488 Better handling of changes to config files
When updating a previous installation from a new release that contains changes
to default/initial configuration files such as environment.conf or
environment.xml, update the user's configuration too if they had not made any
modifications from the previous default file.

If they HAVE made alterations, warn that changes in the default have not been
applied.

Remove the spurious additional call to init code called by the post-uninstall
script, which was interfering with this process.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>