HPCC-9473 TOPN activity crashes if LIMIT depends on context
CHThorTopNActivity queries the LIMIT in the constructor. However there are
cases where that limit is computed in a parent activity, so it must be queried
later.
This fix moves the query, and the allocation of the buffer, from the ctor
to the ready() method, and moves the delete from the dtor to the done method

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>