HPCC-9535 Rationalize Roxie topology info
Ensure that all components use only the information really needed - this will
be backward compatible on all environments that are internally consistent,
i.e. those generated by configmgr. Duplicate nodes in the server list will be
ignored, port information will be taken from RoxieFarmProcess nodes, and
RoxieSlaveProcess and RoxieSlave nodes will be ignored.

This commit does not update configmgr to stop it generating the additional
info (yet).

Remove the creation of the (unused, as far as I can see) 'farm' group for a
roxie when environment is loaded to Dali. All roxie schemes now generate
groups that contain all nodes, just once - as far as I can tell the previous
code would have done so for schemes other than 'full redundancy', which would
have generated a group with most of the servers missing.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>