HPCC-9549 Library cannot be resolved unless uses same package as query
There is a check in the lookupLibrary code that requires that any library
found use the exact same package as the query that is looking it up.

On reflection I think that this check is over-cautious. It's quite likely that
a library will want to use a separate package. Remove the check.

Also remove some unused code in the package interface.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>