HPCC-9560 - ensure spilling occurs if rehash fails
If rehash fails (because no memory available for larger HT),
ensure that spill setup, because the new row being added must
be committed to disk.

At the moment, rehash's inability to allocate the larger HT will
cause all buckets including caller to spill, so it's a non-issue,
but roxiemem may change to avoid calling all callbacks on a alloc
that can't be satisfied.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>