HPCC-9611 Can't run more than one thor job that uses Java
The libjavaembed plugin does not operate correctly if loaded and unloaded
(this is due to JVM limitations).

This does not cause issues for Roxie or hThor, but thor loads referenced
plugins per job, and will thus fail the second time a job references java.

Rather than refactor the (obsolete) plugin mechanism to allow indication of
'keep loaded' status, add code into the plugin itself to 'leak' a reference in
order to prevent it from unloading after having been loaded for the first
time.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>