HPCC-9661 Move column names out of html input tags
For some environment, ECLWatch "Get Roxie Metrics" shows no column
names to select. The existing html code has the column name between
an open INPUT tag and a closing INPUT tag. Since the closing tag
is not standard html, the code of column name may be ignored.
The problem is fixed by moving the column name out of the INPUT tag.

When a column is checked, the column will be added to display.
Another problem found is that the table header is not refreshed
correctly for Chrome. New code is added to detect Chrome and
force the table to refresh as for Firefox.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>