HPCC-9783 - Fix issue with prefix self join, HINT(parallel_match)
The self join helper, doesn't handle prefix joining, so the
regular join helper is used if a prefix join is detected.
However, when HINT(parallel_match) is used and the multi threaded
helper is overlayed, it incorrectly switched some code based on
selfjoin. As a consequence it mismatched and lost records.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>