HPCC-9663 Dynamically start/stop scope scans
Currently, to enable or disable scope scans, the admin has to bring down Dali,
modify the configuration file, and restart dali. This is awkward, and can
cause production to be down resulting in lost revenue.
This change allows scope scan status to be queried, and allows it to be
enabled/disabled dynamically by sending a new request from ESP to Dali in
response to the admin clicking a new button on the permissions page of ECLWatch.
If the configured Dali is downlevel (a version that does not support this feature)
then the button is not displayed

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>