HPCC-9841 Deleting files from cluster other than original fails
If a file is present on more than one cluster, deleting the file from one
cluster is liable to delete the physical files from the original location of
the file rather than any subsequent copy.

Typically this means that when a file is created on thor then copied to Roxie,
an attempt to delete the roxie copy (via eclwatch) is liable to remove the
logical entry for the roxie copy, but the physical files for the thor entry.

The entire mechanism for multiple copies of data on more then one cluster
relies on the cluster descriptors being properly set up with directory
location at all times (but they were often not). This change should ensure
that they are, at least more often.

It may end up expanding the size of information stored in dali about file
locations though.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>