HPCC-9884 Roxie does not copy files from remote if local orphan exists
If there is a local orphan copy of a file, Roxie will fail to resolve the file
rather than copying the file from remote location (as it does if the file is
not present). This tends to leave to queries suspended on only SOME nodes,
which confuses everyone.

Added an option ignoreOrphans, defaulting to true, to treat orphans as if they
were not present (and thus overwrite them with new good copies).

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>