HPCC-9503 Return action results when DFU WU restored
In the existing ECLWatch, FileSpray/WorkunitsAction automatically
redirects webpage to Browse DFU Workunits page after a workunit
action is finished. This fix removes the code for the URL redirect.
The FileSpray/WorkunitsAction returns the action results. It allows
the caller to decide how to use the results and/or redirect to
another page.

Signed-off-by: Kevin Wang <kevin.wang@lexisnexis.com>