HPCC-9919 Fix core accessing uncompressed workunit resources
The technique used to store the size and retrieve the size of stored
resources were not compatible.  As most resources are formatted
with headers designed to support compression this is usually not
noticed.  But resources stored without the headers would fail because
they relied on the size that was stored as an absolute data resource.

Headers will now always be stored with resources whether attempting to
compress them or not.  The size will be retrieved from the headers.

The old size information is still stored as it can be retrieved in
other ways besides the dlopen/dlsym technique used by dllserver.

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>