HPCC-9835 - avoid issues with transaction actions causing crash
If a transaction action throws an exception in the constructor
it caused nasty side-effects, because (prior to this change) the
base of any action caused it to be added to a actions list inside
the transaction that owned it. So when the constroctor failed
it left a linked object in the actions list which was no longer
valid.
This was 1st noticed by a failing rename transaction - the rename
failed because the file being renamed was owned by a super.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>