HPCC-10156 Add WsTopology method ListTargetClusters
The existing TpTargetClusterQuery returns too much information and
can be slow. A new method is requested which only returns cluster name,
type, and whether the cluster is a 'default' target (for jobs like
syntax check, etc) or not. The method is implemented by this fix.

The default target cluster may be specified in esp.xml (see the other
subtask in the parent issue). If not, the isDefault will be set using
the following logic: Loop through all targets. If an hthor is found,
it will be the 'default'. If no hthor, the first thor cluster will be
the 'default'. If no hther and no thor, the first roxie cluster will
be the 'default'.