HPCC-9916 Allow source DALI and ProcessCluster overide in packagemap
Allow daliip and sourceProcessCluster to be specified in the packagemap.
Both can be specified as attributes for the entire packagemap,
or for a specific package, SuperFile or SubFile.

Also no longer clone information for Foreign files.  Roxie always opens
foreign files remotely (as the foreign keyword was intended to imply).

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>