HPCC-10083 - Some slave exceptions detected too late
When a slave aborts and sends an exception, it aborts the
local graph, aborting some barrier points in the process.
As a consequence, the master can proceed beyond the barrier
point without directly knowing what the exception was.
If the timings is unlucky, the master will proceed on to the
next subgraph before being receiving the original exception
from the slave(s).

Change so that the exception is transmitted with the barrier
cancel, so the master can abort with the relevant exception
immediately.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>