HPCC-3315 - Spot cluster 'realCluster' mismatch
If a cluster definition changes it's nodeGroup, then it should
be considered as not matching the previous environment
definition.
And as a consequence it will create that new group when the
environment is loaded, either at Dali restart or updtdalienv time

The consequence for this _not_ happening, is that a nodeGroup
change causes Dali/updtdalienv to think nothing has changed and
to avoid instantiating the group,
causing e.g. Thor to fail to start

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>