HPCC-9521 Refactor CDfsLogicalFileName to make it purely representational
of a logical file, avoid it resolving/looking up

Currently CDfsLogicalFileName resolves any wildcards in logical file names
when it is instantiated. This is unnecessary since in most cases they dont
need to be. This pull request delays the expansion until it is actually
needed, by adding a new "isWild" and "expand" method to the interface.

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>