HPCC-9982 Jobs run on hthor should fail if they need to rebuild a thor persist
Changing clusters in thor jobs doesn;t work, so a persist that names the
cluster it wants to run on will try to rebuild on hthor if a job that uses it
is accidentally run on hthor (or deliberately run on hthor in the hope that
the persist did not need to be rebuilt.

This change only protects cases where a specific cluster has been specified
for a persist. A future commit will look at adding a 'platform type' attribute
on a persist.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>