HPCC-9968 - New EXPIRE semantics
EXPIRE will set # days, not an absolute time as before per file.
It will only set it at all if a flag is generated by the code
generator, however 0 is a valid # of days, which will tell Sasha
to use it's default.

Persist and regular files have a different default # days,
configurable via configmgr.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>