HPCC-10032 Add schema generator support for xpath('')
Add support for xpath('') to schema generation code in both hqlutil.cpp
and fileview2.

For child sets, datasets, and rows, xpath('') needs to generate
schemas where children of the field become direct children of the
current record. For simple types the parent complexType needs to
be marked as containing mixed content.  When xpath('') is nested
across multiple child levels, both mixed content and child elements
need to move up to the nearest parent with out xpath('').

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>