HPCC-10231 - KeyedJoin hitting spurious limits
The end of group marker was encoded in the fpos.
As a consequence a very large fpos could trigger a false limit,
causing the group to be pruned (ATMOST) or a false limit
exception to be fired.
More likely, the index contained no fpos, instead payload fields
occupied that area and as a consequence it's value could also
trigger the false-positive limit.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>