HPCC-10242 Fix preflight not working for roxie server and thor master
The problem is caused by the code added for showing eclserver into target
cluster page. The code is used to pick up server information from
environment settings using a server name. For a thor master and roxie
server, the code is not needed since the cluster tree has been picked up
from environment settings using the process name. And the code logic does
not work for thor master and roxie server. This fix adds a flag used to
skip that code for thor master and roxie server.