HPCC-10278 Create bug report feature not handling multiple logfiles
Currently ws_workunits uses a single MemoryBuffer in the call to add
logfiles to the bug report. But if there are multiple logfiles, the
buffer gets cleared leaving only the last one. This code change dynamically
allocates MemoryBuffers as needed (one for each logfile) and saves them
in a pointerArray, and frees them up once the ZIP file has been created

Signed-off-by: William Whitehead <william.whitehead@lexisnexis.com>