HPCC-10176 Replace "bug" from ZAP report page for legacy WU Details
I made this fix based on the request of HPCC-10291.

The following changes are made for ZAP (bug) report function of legacy
WU Details page: 1. In the WU Details page, the "Bug Report" icon is
removed. A "Z.A.P. Report" button is added beside the "Reset" button.
2. In the Report form, the title is changed from "Bug Report" to
"Report". The label of the submit button is changed from "Create Bug
Report" to "Create Report". The sentences inside the form start with
capital letters.

Signed-off-by: Kevin Wang kevin.wang@lexisnexis.com