HPCC-10255 - Sasha not archiving/restoring associated files
Associated files used to be published as GeneratedDlls, only the query
dll is now published in that way, the rest appear in the workunits
associated file list only.
Add code to archive and restore these too.

There's also an efficiency change here. When a workunit is
archived or restored, it used to connect to /WorkUnits to perform
the locking and to verify the workunit did not already exist.
As a consequence, it would have required the client to pull the
children of /Workunits (all workunits nodes) to the client.
This commit changes it, so that the lock is direct to the
workunit required.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>