HPCC-10345 Disable Banner Setup for non-admin if authentication is set
A flag 'SuperUser' is used to decide whether the web link of Banner Setup
should be displayed or not. (The flag is set to true if no authentication
is set or the esp service is used by an admin user.) The existing code
still shows the Setup link for non admin user due to a bug inside index.xslt
file. The bug is fixed by this fix.

Signed-off-by: Kevin Wang kevin.wang@lexisnexis.com