HPCC-10481 - Default replicate dir causing problems
Set the default replicate directory based on type in clusterinfo.
Was causing problems if not, when other components of different
type than the group type, e.g. processing a thor group file on
eclagent. It would then get local (e.g. eclagent) default dir
and construct the wrong file path.

The symptom seen was that when eclagent deleted files, it failed
to delete replicate parts.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>