HPCC-10302 - Fix eclagent deadlock if fileservices calls in apply
An old change to getPlatform() had the side-effect of causing
calls to updateWorkUnit to deadlock, re-attempting to lock the
workunit already locked by the agent.

This change removes a kludgy way it was checking and casting to
see if a compliant context was available or not.

It also simplies WUmessage to standardize on the already
available addWuException method, which was being called in some
cases already.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>