HPCC-10511 Code using dmetaphone or saltlib may get undefined symbols
In 4.0.4, the ‘plugin proxy’ files would only have been shipped as part of the
clienttools package. As part of the refactoring of the clienttools build, they
appear to be getting included in the regular install too, meaning it’s
indeterminate whether the ‘real’ or the ‘proxy’ definition of dmetaphone and
saltlib gets picked up. If you get the proxy version, you get a link error.

Split the 'header' from the plugin solves the problem.

Signed-off-by: Richard Chapman <rchapman@hpccsystems.com>