HPCC-10482 - Associated files being deleted too early
Archiving is two stage, 1st create the archive and copies,
2nd delete and clearup. Some workunit states are supposed to
prevent the deletion and clearup, e.g. if the query is a scheduled
or a service. However, the archiving code due to an ancient
regression, was deleting associated workunit files too early.
[ For reference, bug introduced by:
https://trac.br.seisint.com/tracdb/changeset/33600 ]

Other problems have been masking the issue,
e.g. HPCC-2558 and HPCC-10255

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>