HPCC-10514 - Spurious boolean, causing premature timeouts
Super file lookups in a couple of the transaction types could
timeout.
An ancient change that removed a boolean, left in a boolean
in a lookupSuperFile call, which was being interpretted as the
timeout of 1ms.
Most of the time it would still get away with it, as the window
was small enough in Dali that 1ms on a the unambiguous superfile
path, meant it succeeded within 1ms.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>