HPCC-10610 ecl-packagemap-add support for specifying a src file prefix
To support accessing some files on 702 systems we need to be able to
specify a prefix to use when looking up the src file, but which will
not be added to the destination file.

The prefix is noted in the file definition attributes as
@cloneFromPrefix.

The prefix may be specified as an extenion to daliips specified either
on the command line or in the packagemap.

Format: --daliip=cluster@ip/prefix or <Package daliip="cluster@ip/prefix">

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexis.com>