HPCC-10594 : Block access to foreign dali from logicalfilelist
Fileservice calls can bypass the local Dali (and therefore the
check from slaves) and go directly to a foreign Dali.
This is highly undesirable, for the same reason it was blocked
by default from the local environment's Dali. It causes a flood
of requests that can usually be recoded more efficiently by
ensuring the request is called once or outside slaves.

LogicalFileList was the only call that in FileServices that
directly called a DFS call that sent a message directly.

Signed-off-by: Jake Smith <jake.smith@lexisnexis.com>