Add architecture macros
These macros are to set the target type to the four known unique
targets: x86, x86_64, ARM32 and ARM64. You should not have more
than one of them set in any given architecture, so if x86 is set,
this is *not* a 64-bit machine.

These macros should replace _64BIT_ and all x86 related ones, when
the _64BIT_ macro means actually x86_64, and help the migration to
ARM platforms, to come.